WIP: cleaning treemapservice, defining correct interfaces, moving method to correct service
